% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_numeric}
\alias{check_numeric}
\title{Check if input is a numeric scalar, vector, or matrix, with no missing values}
\usage{
check_numeric(x, object_type = c("scalar", "vector", "matrix"), finite = TRUE)
}
\arguments{
\item{x}{}

\item{object_type}{}

\item{finite}{}
}
\value{
No returned value
}
\description{
Check if input is a numeric scalar, vector, or matrix, with no missing values
}
