% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wholenumber.R
\name{check_wholenumber}
\alias{check_wholenumber}
\title{Check if input is a whole number (positive integer)}
\usage{
check_wholenumber(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
No return output.
}
\description{
This function checks to see if \code{x} is a positive integer (a whole number).
If \code{x} is not a whole number, the function stops the current R process and
returns an error.
}
