% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E_function.R
\name{E_function}
\alias{E_function}
\title{Calculate stability condition for \eqn{E(f_{a,k}) = \eta}.}
\usage{
E_function(loga, k, B, r, eta)
}
\arguments{
\item{loga}{log(a) value, to optimize over}

\item{k}{integer value between k_min, ..., B}

\item{B}{number of complementary pairs splits}

\item{r}{r-concavity parameter (negative, e.g. -1/2, -1/4)}

\item{eta}{false discovery rate threshold, e.g. \eqn{\theta = q/p} - the target mean for \eqn{E(f_{a,k})}.}
}
\value{
logit-difference between \eqn{E(f_{a,k})} and \eqn{\eta}
}
\description{
See \insertCite{shah_variable_2013}{fastfrechet}, Appendix A4.
}
