% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetreg_mean.R
\name{frechetreg_mean}
\alias{frechetreg_mean}
\title{Fréchet Regression Mean Calculator}
\usage{
frechetreg_mean(
  X,
  Y,
  Z = NULL,
  lambda = NULL,
  lower = -Inf,
  upper = Inf,
  eps = 1e-10,
  skip_checks = FALSE
)
}
\arguments{
\item{X}{A \verb{(n x p)} matrix where all entries are non-\code{NA}, non-\code{NaN}, and finite.}

\item{Y}{A \verb{(n x m)} matrix that is row-wise monotone non-decreasing. This matrix has entries that obey the user-specified box contraints given by the \code{lower} and \code{upper} parameters.}

\item{Z}{An optional \verb{(z x p)} matrix (default \code{NULL}).}

\item{lambda}{An optional \verb{(p x 1)} long vector with non-negative entries and sum across entries is strictly positive.}

\item{lower}{An optional numeric scalar (default \code{-Inf}) that must be strictly less than \code{upper}.}

\item{upper}{An optional numeric scalar (default \code{Inf}) that must be strictly greater than \code{lower}.}

\item{eps}{An optional numeric scalar (default \code{1e-10}) that must be strictly positive.}

\item{skip_checks}{An optional logical argument (default \code{FALSE}). When \code{TRUE} compatibility and dimension checks are run on \code{X}, \code{Y}, \code{Z},\code{lambda}, \code{lower}, \code{upper}, and \code{eps} arguments.}
}
\description{
A short description...
}
