% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRiSO.R
\name{FRiSO}
\alias{FRiSO}
\title{Fréchet Ridge Selection Operator (FRiSO)}
\usage{
FRiSO(
  X,
  Y,
  lower = -Inf,
  upper = Inf,
  tauseq,
  lambda_init = NULL,
  eps = 1e-05,
  nudge = 0,
  alpha = 0.9,
  maxIter = 1000,
  max_theta = pi/4,
  bet = 1
)
}
\arguments{
\item{X}{An n by p matrix column-wise consisting of predictor vectors}

\item{Y}{An n by m matrix row-wise consisting of empirical quantile functions, each evaluated on an equispaced m-grid on (0, 1)}

\item{lower}{Lower box constraint on quantile functions; default is -Inf, i.e. unbounded}

\item{upper}{Upper box constraint on quantile functions; default is +Inf, i.e. unbounded}

\item{tauseq}{A p-length vector containing 'tau' values at which to solve FRiSO problem}

\item{lambda_init}{An optional p-length vector giving the initial 'allowance vector' lambda for FRiSO algorithm; will be scaled to sit on tau-simplex for first entry of tauseq}

\item{eps}{A non-negative error tolerance parameter}

\item{nudge}{A non-negative numeric value to offset warm starts to avoid spurious boundary values}

\item{alpha}{A non-negative dampening parameter (used for 'GSS' and 'SSCG' methods)}

\item{maxIter}{An integer giving the maximum number of iterations for the algorithm to run; default is 1000.}
}
\value{
A p by length(tauseq) matrix column-wise containing fitted 'allowance vectors' lambda per 'tau' in tauseq.
}
\description{
Fréchet Ridge Selection Operator (FRiSO)
}
