% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRiSO_CPSS_univar2wass.R
\name{FRiSO_CPSS_univar2wass}
\alias{FRiSO_CPSS_univar2wass}
\title{Complementary Pairs Stability Selection (CPSS) for FRiSO, for Univariate Distribution Responses Equipped with the 2-Wasserstein Metric}
\usage{
FRiSO_CPSS_univar2wass(
  X,
  Y,
  B = 50,
  thresh = 1e-04,
  new_splits_per_tau = TRUE,
  ...
)
}
\arguments{
\item{X}{An \link{n by p} matrix column-wise consisting of predictor vectors.}

\item{Y}{An \link{n by m} matrix row-wise consisting of empirical quantile functions, each evaluated on a uniformly space m-grid on (0, 1).}

\item{B}{Positive integer number of complementary pair sub-samples to take of original data set (default 50).}

\item{thresh}{A finite positive scalar selection threshold, where \eqn{\hat{lambda}_k(\tau) > }\code{thresh} means the \eqn{k^{th}} variable is selected, and not selected otherwise.}

\item{new_splits_per_tau}{}

\item{...}{Any other required or optional parameters to \code{FRiSO_univar2wass} function, matched only by exact name.}
}
\value{
List containing: 'tau', a vector containing the values of tau which were fed into the function call; 'selected_variables' element, a \link{B by 2 by length(tauseq) by p} array containing 0's and 1's identifying which variables were selected by CPSS per sub-sample; 'selected_samples' element, a \link{B x 2 x length(tauseq) x n} array containing 0's and 1's identifying which samples were selected in course of CPSS; a 'stability_paths' element, a \link{length(tauseq) x p} matrix containing stability measures for each variable (column-wise) against given tauseq.
}
\description{
Complementary Pairs Stability Selection (CPSS) for FRiSO, for Univariate Distribution Responses Equipped with the 2-Wasserstein Metric
}
\examples{
# Generate data for X and Y inputs by using the output of `generate_zinbinom_qf`
n = 100  # number of samples - nrow(X) and nrow(Y).
p = 10   # number of covariates - ncol(X).
m = 100  # EQF grid density - ncol(Y).
lower = 0
upper = Inf

set.seed(31)
mydata = fastfrechet::generate_zinbinom_qf(n = n,
                                           p = p,
                                           m = m)
X = mydata$X  # (n x p) matrix of covariates
Y = mydata$Y  # (n x m) matrix of EQFs, stored row-wise

# Set complementary pairs stability selection parameters
B = 50
thresh = 0.0001
tauseq = 1:10
eps = 0.001

# Run complementary pairs stability selection
cpss = FRiSO_CPSS_univar2wass(X = X,
                              Y = Y,
                              B = B,
                              thresh = thresh,
                              lower = lower,
                              upper = upper,
                              tauseq = tauseq,
                              eps = eps)

# Plot stability paths
matplot(cpss$tau, cpss$stability_paths, type = 'l', lty = 1, lwd = 2,
        col = c(rep('red', 4), rep('black', p - 4)))
}
