% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRiSO_univar2wass.R
\name{FRiSO_univar2wass}
\alias{FRiSO_univar2wass}
\title{Fréchet Ridge Selection Operator (FRiSO)}
\usage{
FRiSO_univar2wass(
  X,
  Y,
  lower = -Inf,
  upper = Inf,
  tauseq,
  lambda_init = NULL,
  eps = 1e-05,
  nudge = 0,
  alpha = 0.9,
  maxIter = 1000,
  max_theta = pi/4,
  impulse = 1
)
}
\arguments{
\item{X}{A \verb{(n x p)} "input" covariate matrix with no missing, all finite entries.}

\item{Y}{A \verb{(n x m)} matrix of observed quantile functions, row-wise monotone non-decreasing. Entries must obey user-specified box constraints given by \code{lower} and \code{upper} parameters.}

\item{lower}{An optional numeric scalar (default \code{-Inf}) lower box constraint; must be strictly less than \code{upper}.}

\item{upper}{An optional numeric scalar (default \code{Inf}) upper box constraint; must be strictly greater than \code{lower}.}

\item{tauseq}{A (\code{p} \eqn{\times} \code{1}) vector containing 'tau' values at which to solve FRiSO problem.}

\item{lambda_init}{An optional (\code{p} \eqn{\times} \code{1}) vector giving the initial 'allowance vector' lambda for FRiSO algorithm (default \code{NULL}); will be scaled to sit on tau-simplex for first entry of \code{tauseq}.}

\item{eps}{A non-negative error tolerance parameter (default \code{1e-5}).}

\item{nudge}{A non-negative numeric scalar to offset warm starts to avoid spurious boundary values (default \code{0}).}

\item{alpha}{A non-negative dampening parameter (default \code{0.9}).}

\item{maxIter}{An integer giving the maximum number of iterations for the algorithm to run (default \code{1000}).}

\item{max_theta}{A step-size scalar parameter no larger than \code{pi / 4} (default \code{pi / 4}).}

\item{impulse}{A scalar between \code{0} and \code{1} which controls the "impulse" in gradient descent with momentum (default \code{1}). \code{impulse} equal to \code{1} means no momentum and \code{impulse <1} means with momentum.}
}
\value{
A (\code{p} \eqn{\times} \code{length(tauseq)}) matrix column-wise containing fitted 'allowance vectors' lambda per 'tau' in tauseq.
}
\description{
A short description...
}
