% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_zinbinom_qf.R
\name{generate_zinbinom_qf}
\alias{generate_zinbinom_qf}
\title{Title}
\usage{
generate_zinbinom_qf(n, p, m, zero_inflation = 0.2, prob = 0.5, size = 10)
}
\arguments{
\item{n}{}

\item{p}{}

\item{m}{}

\item{zero_inflation}{}

\item{prob}{}

\item{size}{}
}
\description{
Title
}
