% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minD.R
\name{minD}
\alias{minD}
\title{Calculate min(D...) values.}
\usage{
minD(theta, B, r = c(-1/2, -1/4))
}
\arguments{
\item{theta}{false discovery rate threshold}

\item{B}{number of complementary pairs splits}

\item{r}{r-concavity parameter (negative, e.g. -1/2, -1/4)}
}
\value{
the solution to min(D...) from Shah and Samworth (2013), eq. 8.
}
\description{
See \insertCite{shah_variable_2013}{fastfrechet}, eq. 8.
}
