% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shah_Samworth_thresholds.R
\name{Shah_Samworth_thresholds}
\alias{Shah_Samworth_thresholds}
\title{Pointwise Selection Thresholds for Complementary Pairs Stability Selection (CPSS)}
\usage{
Shah_Samworth_thresholds(p, q, B = 50, E_thr = 1)
}
\arguments{
\item{p}{A positive integer giving the total number of variables.}

\item{q}{A numeric vector giving average model size(s) after variable selection, each entry strictly within \verb{(0, p)}.}

\item{B}{A positive integer giving the number of complementary pairs splits (default \code{50}).}

\item{E_thr}{A positive numeric scalar giving the desired constraint on the number of final selected variables with "low selection probability" (default \code{1}).}
}
\value{
A list object with components:
\tabular{ll}{
\code{E_thr} \tab returns the selection constraint, as provided to the function. \cr
\code{B} \tab returns the number of complementary pairs splits, as provided to the function. \cr
\code{relative_model_size} \tab returns the relative model size, i.e. \code{q / p} from the function inputs. \cr
\code{pointwise_thresholds} \tab returns a \code{p}-long vector of calculated selection thresholds. See details. \cr
}
}
\description{
This function calculates pointwise selection thresholds for CPSS from
\insertCite{shah_variable_2013}{fastfrechet}. Empirical CPSS selection proportions
above these thresholds have pointwise control on the number of selected variables
which have "low selection probability" \eqn{\theta}. This function uses the default
\eqn{\theta = q/p}, where \code{q} is the estimated model size after variable selection.
}
\details{
CPSS obtains a final selected model by identifying variables selected a high
proportion of times during repeated data splits. It is a meta-procedure that
works independent of the baseline variable selection procedure, such as LASSO
or, in the case of variable selection for Fréchet regression, FRiSO. Under
certain conditions, \insertCite{shah_variable_2013}{fastfrechet} derived
thresholds for this empirical selection proportion that control the expected
number of variables which have "low selection probability", i.e. are disfavored
by the baseline procedure like FRiSO. This function inverts the problem by
identifying the minimum threshold necessary to control the selection procedure.

\insertCite{shah_variable_2013}{fastfrechet} recommend defining "low selection
probability" as relative to model size. A variable has low selection probability
if it is selected less frequently than the average relative model size, that
is, the number of selected variables \code{q} divided by the number of total
variables \code{p}. This function takes \code{p} and \code{q} as inputs (\code{q} may be a vector),
as well as the number of CPSS data splits \code{B} and desired control level \code{E_thr}.
Note that error control is pointwise per \code{q}.
}
\examples{
# Specify number of parameters and vector of post-selection model sizes:
p <- 30
q <- 1:10

# Specify number of CPSS splits, and desired error control:
B <- 50
E_thr <- 1

# Calculate thresholds:
output <- Shah_Samworth_thresholds(p = p, q = q, B = B, E_thr = E_thr)

# Plot calculated thresholds against relative model sizes:
plot(output$relative_model_size, output$pointwise_thresholds,
  type = "o", pch = 20, xlab = expression(theta),
  ylab = expression(pi * "(" * theta * ")"),
  main = "Shah and Samworth Thresholds", ylim = c(0, 1)
)
}
\references{
\insertRef{shah_variable_2013}{fastfrechet}
}
