% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_function.R
\name{T_function}
\alias{T_function}
\title{Calculate \eqn{T_t(g_{a,k})}.}
\usage{
T_function(a, k, B, r, eta, t)
}
\arguments{
\item{a}{a value, to optimize over}

\item{k}{integer value between k_min, ..., B}

\item{B}{number of complementary pairs splits}

\item{r}{r-concavity parameter (negative, e.g. -1/2, -1/4)}

\item{eta}{false discovery rate threshold}

\item{t}{candidate variable selection threshold}
}
\value{
calculated \eqn{T_t(g_{a,k})} value, eq. (17) in Appendix A4 of
Shah and Samworth (2013)
}
\description{
See \insertCite{shah_variable_2013}{fastfrechet}, Appendix A4.
}
