% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_prob.R
\name{tail_prob}
\alias{tail_prob}
\title{Calculate tail probabilities for r-concave functions.}
\usage{
tail_prob(B, eta, r)
}
\arguments{
\item{B}{number of complementary pairs splits}

\item{eta}{false discovery rate threshold}

\item{r}{r-concavity parameter (negative, e.g. -1/2, -1/4)}
}
\value{
a list containing thresholds and calculated tail probabilities
}
\description{
See \insertCite{shah_variable_2013}{fastfrechet}, Appendix A4.
}
