% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wholenumber.R
\name{check_wholenumber}
\alias{check_wholenumber}
\title{Check if input is a whole number (positive integer)}
\usage{
check_wholenumber(x)
}
\arguments{
\item{x}{}
}
\value{
No return output.
}
\description{
Check if input is a whole number (positive integer)
}
