% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetreg_univar2wass.R
\name{frechetreg_univar2wass}
\alias{frechetreg_univar2wass}
\title{(Global) Fréchet Regression for Univariate Distributions with 2-Wasserstein Metric}
\usage{
frechetreg_univar2wass(
  X,
  Y,
  Z = NULL,
  lambda = NULL,
  lower = -Inf,
  upper = Inf,
  eps = 1e-10,
  skip_checks = FALSE
)
}
\arguments{
\item{X}{A \verb{(n x p)} "input" covariate matrix with no missing, all finite entries.}

\item{Y}{A \verb{(n x m)} matrix of observed quantile functions, row-wise monotone non-decreasing. Entries must obey user-specified box constraints given by \code{lower} and \code{upper} parameters.}

\item{Z}{An optional \verb{(z x p)} "output" covariate matrix (default \code{NULL}) with no missing, all finite entries.}

\item{lambda}{An optional \verb{(p x 1)} vector with non-negative entries whose sum is strictly positive.}

\item{lower}{An optional numeric scalar (default \code{-Inf}) lower box constraint; must be strictly less than \code{upper}.}

\item{upper}{An optional numeric scalar (default \code{Inf}) upper box constraint; must be strictly greater than \code{lower}.}

\item{eps}{An optional numeric scalar (default \code{1e-10}) error tolerance; must be strictly positive.}

\item{skip_checks}{An optional logical argument (default \code{FALSE}). When \code{TRUE} compatibility and dimension checks are skipped.}
}
\description{
This function calculates global Fréchet regression mean objects (Petersen et al. 2019), for the space of univariate distribution responses equipped with the 2-Wasserstein metric. Observed distributions are assumed to be quantile functions evaluated on a common, equally spaced \code{m}-grid in (0, 1). Options for user-specified output covariate matrix, generalized ridge sparsity vector (Tucker et al. 2023), and box constraints. The Fréchet regression problem in this context is reducible to a quadratic programming problem; the workhorse of this function is a custom active set method based on Arnström et al. 2022.
}
