% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetreg_univar2wass.R
\name{frechetreg_univar2wass}
\alias{frechetreg_univar2wass}
\title{(Global) Fréchet regression for univariate distributions with 2-Wasserstein metric}
\usage{
frechetreg_univar2wass(
  X,
  Y,
  Z = NULL,
  C_init = NULL,
  lambda = NULL,
  lower = -Inf,
  upper = Inf,
  eps = 1e-10
)
}
\arguments{
\item{X}{A (\code{n} \eqn{\times} \code{p}) "input" covariate matrix with no missing, all finite entries.}

\item{Y}{A (\code{n} \eqn{\times} \code{m}) matrix of observed quantile functions evaluated on a shared uniform \code{m}-grid in \eqn{(0, 1)} - e.g. \code{seq(0.5/m, 1 - 0.5/m, len = m)} - row-wise monotone non-decreasing. Entries must obey user-specified box constraints given by \code{lower} and \code{upper} parameters.}

\item{Z}{An optional (\code{z} \eqn{\times} \code{p}) "output" covariate matrix (default \code{NULL}) with no missing, all finite entries.}

\item{C_init}{An optional (\code{n} \eqn{\times} \code{m + 1}) matrix (or \code{z} \eqn{\times} \code{m + 1} if \code{Z} is provided) of non-negative entries, specifying initial active set(s) for optimization. Active sets are identified by positive entries, row-wise if a matrix.}

\item{lambda}{An optional (\code{p} \eqn{\times} \code{1}) vector with non-negative entries whose sum is strictly positive.
If \code{NULL} a non-regularized regression is performed (the default).}

\item{lower}{An optional numeric scalar (default \code{-Inf}) lower box constraint; must be strictly less than \code{upper}.}

\item{upper}{An optional numeric scalar (default \code{Inf}) upper box constraint; must be strictly greater than \code{lower}.}

\item{eps}{An optional numeric scalar (default \code{1e-10}) error tolerance; must be strictly positive.}
}
\value{
A list object with components:
\tabular{ll}{
\code{Qhat} \tab returns a (\code{n} \eqn{\times} \code{m}) matrix - or (\code{z} \eqn{\times} \code{m}) matrix if \code{Z} was provided - row-wise containing the solutions to the Fréchet regression problem, i.e. best-fitting quantile functions bounded between \code{lower} and \code{upper}. These quantile functions are treated as evaluated on a shared \code{m}-grid in \eqn{(0, 1)}, e.g. \code{seq(0.5/m, 1 - 0.5/m, len = m)}.\cr
\code{Lagrange_Multiplier} \tab returns a (\code{n} \eqn{\times} (\code{m + 1})) matrix - or (\code{z} \eqn{\times} (\code{m + 1})) matrix if \code{Z} was provided - row-wise containing the Lagrange multipliers associated with the underlying QP problems. \cr
}
}
\description{
Calculate global Fréchet regression mean objects \insertCite{petersen_frechet_2019}{fastfrechet},
for the space of univariate distribution responses equipped
with the 2-Wasserstein metric. Observed distributions are assumed
to be quantile functions evaluated on a common, equally spaced \code{m}-grid
in \eqn{(0, 1)}. Default is to calculate Fréchet regression means conditioned
on input covariate row-vectors, but can calculate Fréchet means on different
set of covariate vectors. Accommodates the regularization scheme from
\insertCite{tucker_variable_2023}{fastfrechet}.
}
\details{
Fréchet regression generalizes Euclidean regression to the general metric space setting.
In the sample setting, we observe covariate-response pairs \eqn{\{(\mathbf{x}_i, \mathbf{y}_i)\} \subset \mathbb{R}^p \times \Omega},
where \eqn{\mathbf{x}_i}'s are rows of matrix \eqn{\mathbf{X}} column-centered
and scaled such that \eqn{\mathrm{diag}(\mathbf{X}^{\top}\mathbf{X}) = \pmb{1}},
and \eqn{\Omega} is the space of univariate quantile functions equipped with
the 2-Wasserstein metric \eqn{d_W(\mathbf{q}, \mathbf{p}) = \lVert\mathbf{q} - \mathbf{p}\rVert_{L_2[0,1]}}.
(Generally we also only observe the distributions over a discrete grid, not as
full functions.) The conditional Fréchet mean is estimated by
\deqn{\widehat{\mathbf{q}}_{\oplus}(\mathbf{x}_i) = \underset{\mathbf{q}\in\Omega}{\mathrm{argmin}} \sum_{j=1}^n \left\{ n^{-1} + \mathbf{x}_i^{\top}(\mathbf{X}^{\top}\mathbf{X})^+\mathbf{x}_j\right\} d_W^2(\mathbf{y}_j, \mathbf{q}).}
This is the Fréchet regression problem, a weighted-mean and quadratic programming
problem which this function solves with a customized dual active-set method
inspired by \insertCite{arnstrom_dual_2022}{fastfrechet}.

Options include box constraints on distribution support, an option to evaluate
quantile function estimates on a different set of covariate vectors than the
"input" vectors, and an option to include an initial estimate of the active
constraint sets. The function accommodates non-full rank covariate matrices,
as well as the regularization used in the variable selection method of
\insertCite{tucker_variable_2023}{fastfrechet} through the parameter \code{lambda}.
By default, the function will perform non-regularized regression.
}
\examples{
# Generate data for X and Y inputs by using the output of `generate_zinbinom_qf`
n <- 100 # number of samples - nrow(X) and nrow(Y).
p <- 10 # number of covariates - ncol(X).
m <- 100 # EQF grid density - ncol(Y).
mseq <- seq(1 / (2 * m), 1 - 1 / (2 * m), length.out = m)

set.seed(31)
mydata <- generate_zinbinom_qf(n = n, p = p, m = m)

X <- mydata$X # (n x p) matrix of covariates
Y <- mydata$Y # (n x m) matrix of EQFs, stored row-wise

# Estimate conditional QFs:
output <- frechetreg_univar2wass(
  X = X,
  Y = Y,
  Z = NULL,
  C_init = NULL,
  lambda = NULL,
  lower = 0,
  upper = Inf
)

# Note: to numerical precision, these QFs are non-decreasing...
min(apply(output$Qhat, 1, diff))

# ...and bounded from below by the lower bound, zero:
min(output$Qhat)

# Plot the conditional QFs:
plot(
  x = c(), y = c(), xlim = c(0, 1), ylim = c(0, max(output$Qhat)),
  main = "Fréchet Regression QFs", xlab = "p", ylab = "quantile"
)
for (i in 1:n) lines(mseq, output$Qhat[i, ], lwd = 2)
}
\references{
\insertRef{arnstrom_dual_2022}{fastfrechet}

\insertRef{petersen_frechet_2019}{fastfrechet}

\insertRef{tucker_variable_2023}{fastfrechet}
}
