% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frechetreg_univar2wass.R
\name{frechetreg_univar2wass}
\alias{frechetreg_univar2wass}
\title{(Global) Fréchet Regression for Univariate Distributions with 2-Wasserstein Metric}
\usage{
frechetreg_univar2wass(
  X,
  Y,
  Z = NULL,
  C_init = NULL,
  lambda = NULL,
  lower = -Inf,
  upper = Inf,
  eps = 1e-10
)
}
\arguments{
\item{X}{A (\code{n} \eqn{\times} \code{p}) "input" covariate matrix with no missing, all finite entries.}

\item{Y}{A (\code{n} \eqn{\times} \code{m}) matrix of observed quantile functions, row-wise monotone non-decreasing.
Entries must obey user-specified box constraints given by \code{lower} and \code{upper} parameters.}

\item{Z}{An optional (\code{z} \eqn{\times} \code{p}) "output" covariate matrix (default \code{NULL}) with no missing, all finite entries.}

\item{C_init}{An optional (\code{n} \eqn{\times} \code{m + 1}) matrix (or \code{z} \eqn{\times} \code{m + 1} if \code{Z} is provided) of non-negative entries, specifying initial active set(s) for optimization. Active sets are identified by positive entries, row-wise if a matrix.}

\item{lambda}{An optional (\code{p} \eqn{\times} \code{1}) vector with non-negative entries whose sum is strictly positive.
If \code{NULL} a non-regularized regression is performed (the default).}

\item{lower}{An optional numeric scalar (default \code{-Inf}) lower box constraint; must be strictly less than \code{upper}.}

\item{upper}{An optional numeric scalar (default \code{Inf}) upper box constraint; must be strictly greater than \code{lower}.}

\item{eps}{An optional numeric scalar (default \code{1e-10}) error tolerance; must be strictly positive.}
}
\value{
A (\code{n} \eqn{\times} \code{m}) matrix that is the unique solution to the Fréchet Regression problem for univariate distribution responses.
The solution has rows that are monotone non-decreasing and bounded between the \code{lower} and \code{upper} arguments.
}
\description{
\code{frechetreg_univar2wass()} calculates global Fréchet regression mean
objects \insertCite{petersen_frechet_2019}{fastfrechet},
for the space of univariate distribution responses equipped
with the 2-Wasserstein metric. Observed distributions are assumed
to be quantile functions evaluated on a common, equally spaced \code{m}-grid
in \eqn{(0, 1)}. Options include user-specified output covariate matrix,
generalized ridge sparsity vector \insertCite{tucker_variable_2023}{fastfrechet},
and box constraints. The Fréchet regression problem in this context is reducible
to a quadratic programming problem; the workhorse of this function is a custom
active set method based on \insertCite{arnstrom_dual_2022}{fastfrechet}.
}
\examples{
# Generate data for X and Y inputs by using the output of `generate_zinbinom_qf`
n = 100  # number of samples - nrow(X) and nrow(Y).
p = 10   # number of covariates - ncol(X).
m = 100  # EQF grid density - ncol(Y).
mseq = seq(1 / (2 * m), 1 - 1 / (2 * m), length.out = m)

set.seed(31)
mydata = fastfrechet::generate_zinbinom_qf(n = n,
                                           p = p,
                                           m = m)

X = mydata$X  # (n x p) matrix of covariates
Y = mydata$Y  # (n x m) matrix of EQFs, stored row-wise

# Estimate conditional QFs:
output = fastfrechet::frechetreg_univar2wass(X = X,
                                             Y = Y,
                                             Z = NULL,
                                             C_init = NULL,
                                             lambda = NULL,
                                             lower = 0,
                                             upper = Inf)

# Note: to numerical precision, these QFs are non-decreasing...
min(apply(output$Qhat, 1, diff))

# ...and bounded from below by the lower bound, zero:
min(output$Qhat)

# Plot the conditional QFs:
plot(x = c(), y = c(), xlim = c(0, 1), ylim = c(0, max(output$Qhat)),
     main = "Fréchet Regression QFs", xlab = "p", ylab = "quantile")
for(i in 1:n) lines(mseq, output$Qhat[i, ], lwd = 2)
}
\references{
\insertRef{arnstrom_dual_2022}{fastfrechet}

\insertRef{petersen_frechet_2019}{fastfrechet}

\insertRef{tucker_variable_2023}{fastfrechet}
}
