% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRiSO_CV_univar2wass.R
\name{FRiSO_CV_univar2wass}
\alias{FRiSO_CV_univar2wass}
\title{FRiSO_CV_univar2wass}
\usage{
FRiSO_CV_univar2wass(X, Y, K = NULL, thresh = 1e-04, ...)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{K}{Numeric whole number no bigger than \code{nrow(X)}.}

\item{thresh}{Positive scalar that determines the selection cutoff for the \code{lambda} vector in the variable selection problem.}

\item{...}{other inputs to FRiSO see \code{\link[=FRiSO_univar2wass]{FRiSO_univar2wass()}}.}
}
\value{
A list object with components:
\tabular{ll}{
\code{tauseq} \tab returns a (\code{n} \eqn{\times} \code{m}) matrix that is the unique solution to the Fréchet Regression problem for univariate distribution responses.
The solution has rows that are monotone non-decreasing and bounded between the \code{lower} and \code{upper} arguments. \cr
\code{errors} \tab returns a (\code{n} \eqn{\times} \code{m}) matrix that contains the refitted test errors.  \cr
\code{error_sum} \tab returns a \code{n} length vector that contains the aggregate CV error for each \eqn{\tau}. \cr
\code{opt_tau} \tab returns an (ADD objecttype and dimensions) the \tau that contains \eqn{\tau} that minimizes aggregate CV error. \cr
\code{opt_lambda} \tab returns an (ADD objecttype and dimensions) high-precision 'allowance vector'. \cr
\code{opt_selected} \tab returns an (ADD objecttype and dimensions) Identify selected variables at opt_tau, using given threshold.\cr
}
}
\description{
FRiSO_CV_univar2wass
}
