% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRiSO_CV_univar2wass.R
\name{FRiSO_CV_univar2wass}
\alias{FRiSO_CV_univar2wass}
\title{FRiSO_CV_univar2wass}
\usage{
FRiSO_CV_univar2wass(X, Y, K = NULL, thresh = 1e-04, ...)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{K}{Numeric whole number no bigger than \code{nrow(X)}.}

\item{thresh}{Positive scalar that determines the selection cutoff for the \code{lambda} vector in the variable selection problem.}

\item{...}{other inputs to FRiSO see \code{\link[=FRiSO_univar2wass]{FRiSO_univar2wass()}}.}
}
\value{
A list object with components:
\tabular{ll}{
\code{tauseq} \tab returns the numeric vector containing \eqn{\tau} values at which FRiSO problem was solved. \cr
\code{errors} \tab returns a (\code{p} \eqn{\times} \code{length(tauseq)}) matrix that contains the refitted test errors.  \cr
\code{error_sum} \tab returns a \code{length(tauseq)}-long vector that contains the aggregate CV error for each \eqn{\tau}. \cr
\code{opt_tau} \tab returns a numeric scalar that is the \eqn{\tau} that minimizes the aggregate CV error. \cr
\code{opt_lambda} \tab returns a \code{p}-long numeric vector that is the high-precision 'allowance vector'. \cr
\code{opt_selected} \tab returns a numeric vector (up to \code{p}-long) that contains the indices of the variables selected in the optimal model. \cr
}
}
\description{
FRiSO_CV_univar2wass
}
