% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRiSO_CV_univar2wass.R
\name{FRiSO_CV_univar2wass}
\alias{FRiSO_CV_univar2wass}
\title{FRiSO_CV_univar2wass}
\usage{
FRiSO_CV_univar2wass(X, Y, K = NULL, thresh = 1e-04, ...)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{K}{}

\item{thresh}{}

\item{...}{other inputs to FRiSO}
}
\description{
FRiSO_CV_univar2wass
}
