% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRiSO_univar2wass.R
\name{FRiSO_univar2wass}
\alias{FRiSO_univar2wass}
\title{Fréchet Ridge Selection Operator (FRiSO)}
\usage{
FRiSO_univar2wass(
  X,
  Y,
  lower = -Inf,
  upper = Inf,
  tauseq,
  lambda_init = NULL,
  eps = 1e-05,
  nudge = 0,
  alpha = 0.9,
  maxIter = 1000,
  max_theta = pi/4,
  bet = 1
)
}
\arguments{
\item{X}{A \verb{(n x p)} "input" covariate matrix with no missing, all finite entries.}

\item{Y}{A \verb{(n x m)} matrix of observed quantile functions, row-wise monotone non-decreasing. Entries must obey user-specified box constraints given by \code{lower} and \code{upper} parameters.}

\item{lower}{An optional numeric scalar (default \code{-Inf}) lower box constraint; must be strictly less than \code{upper}.}

\item{upper}{An optional numeric scalar (default \code{Inf}) upper box constraint; must be strictly greater than \code{lower}.}

\item{tauseq}{A p-length vector containing 'tau' values at which to solve FRiSO problem}

\item{lambda_init}{An optional p-length vector giving the initial 'allowance vector' lambda for FRiSO algorithm; will be scaled to sit on tau-simplex for first entry of tauseq}

\item{eps}{A non-negative error tolerance parameter}

\item{nudge}{A non-negative numeric value to offset warm starts to avoid spurious boundary values}

\item{alpha}{A non-negative dampening parameter (used for 'GSS' and 'SSCG' methods)}

\item{maxIter}{An integer giving the maximum number of iterations for the algorithm to run; default is 1000.}

\item{max_theta}{}

\item{bet}{}
}
\value{
A p by length(tauseq) matrix column-wise containing fitted 'allowance vectors' lambda per 'tau' in tauseq.
}
\description{
Fréchet Ridge Selection Operator (FRiSO)
}
