% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_zinbinom_qf.R
\name{generate_zinbinom_qf}
\alias{generate_zinbinom_qf}
\title{Generate Random Zero-Inflated Negative Binomial Quantile Functions}
\usage{
generate_zinbinom_qf(n, p, m, zero_inflation = 0.2, prob = 0.5, size = 10)
}
\arguments{
\item{n}{A positive integer giving the sample size (i.e. rows in the covariate matrix and response matrix).}

\item{p}{A positive integer greater than or equal to 4 giving the number of covariates.}

\item{m}{A positive integer giving the grid density of the quantile functions in \eqn{(0, 1)}.}

\item{zero_inflation}{A numeric value in the range \eqn{[0, 1]} giving the zero inflation parameter of the \code{zinbinom} distribution.}

\item{prob}{A numeric scalar in the range \eqn{(0,1]} giving the probability parameter of the \code{nbinom} part.}

\item{size}{A positive numeric scalar giving the size parameter of the \code{nbinom} part.}
}
\value{
A list object with components:
\tabular{ll}{
\code{X} \tab returns a \verb{(n} \eqn{\times} \verb{p)} covariate matrix. \cr
\code{Y} \tab returns an \verb{(n} \eqn{\times} \verb{m)} quantile function matrix. \cr
}
}
\description{
This function generates random zero-inflated negative binomial (\code{zinbinom})
quantile functions, as functions of covariates \eqn{X_1}, \eqn{X_2}, \eqn{X_3}, and \eqn{X_4}.
Quantile function matrix \code{Y} and covariate matrix \code{X} are generated as in Experiment B in
\insertCite{coulter_fast_2024}{fastfrechet}, and are evaluated on a common,
equally spaced \code{m}-grid in \eqn{(0, 1)}. Options include number of samples
(covariate vectors, and response quantile functions), number of total
covariates (at least 4; generated iid from \eqn{N(0, 1)}), grid density \code{m},
and baseline \code{zinbinom} parameters.
}
\examples{
n <- 100 # number of samples - nrow(X) and nrow(Y).
p <- 10 # number of covariates - ncol(X).
m <- 100 # EQF grid density - ncol(Y).
mseq <- seq(1 / (2 * m), 1 - 1 / (2 * m), length.out = m)

set.seed(31)
mydata <- generate_zinbinom_qf(n = n, p = p, m = m)

X <- mydata$X # (n x p) matrix of covariates
Y <- mydata$Y # (n x m) matrix of EQFs, stored row-wise

# Plot the EQFs:
plot(
  x = c(), y = c(), xlim = c(0, 1), ylim = c(0, max(Y)),
  main = "Zero-Inflated Negative Binomial QFs",
  xlab = "p", ylab = "quantile"
)
for (i in 1:n) lines(mseq, Y[i, ], lwd = 2)
}
\references{
\insertRef{coulter_fast_2024}{fastfrechet}
}
