% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_zinbinom_qf.R
\name{generate_zinbinom_qf}
\alias{generate_zinbinom_qf}
\title{Generate Random Zero-Inflated Negative Binomial Quantile Functions}
\usage{
generate_zinbinom_qf(n, p, m, zero_inflation = 0.2, prob = 0.5, size = 10)
}
\arguments{
\item{n}{A numeric value strictly greater than 0. Determines the number of
rows in the covariate matrix.}

\item{p}{A numeric value strictly greater than 4. Determines the number of
columns in the covariate matrix.}

\item{m}{A numeric value strictly greater than 0.}

\item{zero_inflation}{A numeric value in the range (0,1).}

\item{prob}{A numeric value in the range (0,1).}

\item{size}{A numeric value strictly greater than 0.}
}
\description{
This function generates random zero-inflated negative binomial (zinbinom)
quantile functions, as functions of covariates X1, X2, X3, and X4.
Quantile functions are generated from X parameters as in Experiment B in
\insertCite{coulter_fast_2024}{fastfrechet}, and are evaluated on a common,
equally spaced \code{m}-grid in (0, 1). Options include number of samples
(covariate vectors, and response quantile functions), number of total
covariates (at least 4; generated iid from N(0, 1)), grid density \code{m},
and baseline zinbinom parameters.
}
\references{
\insertRef{coulter_fast_2024}{fastfrechet}
}
